requirements.txt
numpy>=1.19.0
pandas>=1.2.0
scipy>=1.6.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
